+++ examples/node/utils.ts
@@ -1,28 +1,38 @@
 import { type CasinoChain } from "@betswirl/sdk-core";
 import { createConfig, webSocket } from "@wagmi/core";
-import { http, type Hex, createWalletClient } from "viem";
+import { http, type Hex, createWalletClient, createPublicClient } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
 
 export function getWagmiConfigFromCasinoChain(casinoChain: CasinoChain) {
+
+  return createConfig({
+    chains: [casinoChain.viemChain],
+
+    client: () => {
+      return getViemClientsFromCasinoChain(casinoChain).walletClient;
+    },
+  });
+}
+
+export function getViemClientsFromCasinoChain(casinoChain: CasinoChain) {
   const privateKey = process.env.PRIVATE_KEY;
   const rpcUrl = process.env.RPC_URL;
-
   const account = privateKeyToAccount(privateKey as Hex);
   const transport = rpcUrl?.startsWith("wss")
     ? webSocket(rpcUrl)
     : http(rpcUrl);
-  return createConfig({
-    chains: [casinoChain.viemChain],
-
-    client: ({ chain }) => {
-      return createWalletClient({
-        chain,
-        transport,
-        account,
-      });
-    },
+  const walletClient = createWalletClient({
+    chain: casinoChain.viemChain,
+    transport,
+    account,
+  })
+  const publicClient = createPublicClient({
+    chain: casinoChain.viemChain,
+    transport,
   });
+  return { walletClient, publicClient };
 }
+
 
 export function checkEnvVariables() {
   if (!process.env.PRIVATE_KEY) {