+++ packages/core/src/provider/index.ts
@@ -1,2 +1,4 @@
 export * from "./client";
 export * from "./wallet";
+export * from "./viemClient";
+export * from "./viemWallet";