+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @betswirl/sdk-core
 
+## 0.0.3
+
+### Patch Changes
+
+- Add Viem client and wallet
+
 ## 0.0.2
 
 ### Patch Changes